requirements.txt
# AutoMinds Email Assistant - Dependencies
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Google Gmail API
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0

# Microsoft Outlook (Graph API)
msal==1.31.1
httpx==0.28.1

# AI
anthropic==0.44.0

# Database
supabase==2.11.0

# Scheduling
apscheduler==3.10.4

# Security
cryptography==44.0.0
python-jose[cryptography]==3.3.0

# Utilities
jinja2==3.1.5
python-multipart==0.0.18
